docker: Docker image for submitter and reporter. Updated instructions (#14)
This PR significantly revamps the Docker functionality of the vigilante
by adding functionality that enables the automatic deployment either
locally through docker-compose or in the cloud. This is accomplished
through the usage of various configuration files that are attached to
the Docker instance and allow the vigilante to access Babylon and BTC.

The instructions for running the Vigilante have been revamped and the
Docker images for running the reporter and the submitter are now split.

Also, there were some minor bugs regarding a mismatch about the
`mapstructure` attributes and the `sample-vigilante.yaml` configuration
files.