Add instructions for creating a wallet and corresponding config entries (#34)
The submitter needs access to a wallet to perform its operations. This
PR includes wallet related entries in the configuration file:
- A wallet name
- A wallet password (the one that unlocks it)
- The endpoint where the wallet is running
- The CA file that the wallet uses (different than the one used by
btcd).

Also, instructions have been added for creating a wallet on the README.

Important note: This PR changes the port in which btcd is running from
`18554` to `18556` (the default one). This change happened because the
default port for the wallet is `18554` and I don't see a reason why we
should not use the default ports. However, local configuration files
should be changed after this PR gets merged (if someone follows the
README instructions).

Future improvement: Clean up the configuration file a little bit to
separate btcd entries and wallet entries.