reporter: handle re-orgs (#84)
Fixes [BM-210](https://babylon-chain.atlassian.net/browse/BM-210)

This PR adds functionality in the reporter to handle re-orgs. The
following cases are handled

1. Connected event
a. If the connected block parent is not equal to the cache tip, then the
cache is corrupted - restart bootstrapping
    b. else normally submit headers and checkpoint data to BBN
    
2. Disconnected event
a. Disconnected block hash should be the same as cache tip hash. If not,
it means the cache was corrupted - restart
        bootstrapping 
    b. else remove the last element in the cache. 

3. Add RW lock in cache for safe concurrent access. 

## Note
This PR adds re-org handling only in the Btcd subscription model, for
ZMQ I will open a separate PR.