bstracker: metrics for atomic slasher (#52)
This PR adds basic application metrics for the atomic slasher routine
under BTC staking tracker.

It tracks 

- the number of tracked BTC delegations, and
- ~~the number of slashed finality providers due to selective slashing~~

~~Note that metrics about the slashed finality providers/delegations are
provided in the slasher routine's metrics. Atomic slasher sends the
to-be-slashed finality providers to the slasher routine for slashing, so
slasher routine will record these metrics alraedy.~~

What we can do next is to define and implement performance-related
metrics for the vigilante programs.