reporter: handle the gap and overlap between bootstrapping and ongoing execution (#202)
Fixes #201 

This PR improves the flow of bootstrapping and handling ongoing BTC
blocks in vigilante reporter, including

- Added an e2e tests ensuring that when BTC headers are generated
frequently (such that new BTC headers are generated during
bootstrapping), reporter can still ensure consistency between Babylon
and BTC. This also helped reproducing #201.
- Handling events that involve BTC headers overlapped with BTC cache. In
particular, if the BTC header is no higher than the BTC cache tip, then
ignore it. This is possible when bootstrapping is repeated due to fork
issues, and is actually the root cause of #201. This is evidenced by the
failed
[CI](https://app.circleci.com/pipelines/github/babylonchain/vigilante/846/workflows/ba47d802-20df-49ca-b130-bf547ab04e15/jobs/977)
where a lot of such `cache is not up-to-date` errors occur. After this
PR the errors do not occur anymore in e2e.

TODO:
- [ ] reporter still cannot BTC block interval shorter than the latency
of submitting a Babylon tx (a couple of seconds). This needs some
breaking changes in the reporter (e.g., batch submit BTC headers or
submit txs async to Babylon).